Rough design of topbar, restructured files
