+++ frontend/src/Pages/About.tsx
@@ -1,13 +1,9 @@
 import React from "react";
-import logo from "../logo.svg";
 import "../styles/App.css";
 
 function About() {
 	return (
 		<div className="App">
-			<header className="App-header">
-				<img src={logo} className="App-logo" alt="logo" />
-			</header>
 			<h1>
 				<title>About Our Company</title>
 				<link></link>
