+++ frontend/src/Pages/App.tsx
@@ -1,12 +1,10 @@
 import React from "react";
-import logo from "./logo.svg";
-import "./styles/App.css";
+import "../styles/App.css";
 
 function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
         <ul>
           <li>Ben</li>
           <li>Ahsaas</li>
@@ -15,14 +13,6 @@ function App() {
           <li>Ling</li>
           <li>Brandon</li>
         </ul>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
       </header>
     </div>
   );
