+++ frontend/src/styles/TopBar.css
@@ -1,0 +1,14 @@
+.MuiToolbar-root {
+	height: 90px;
+	background-color: azure;
+	justify-content: center;
+}
+
+.MuiButton-root {
+	border-radius: 30px !important;
+	max-height: 30px;
+}
+
+.logo {
+	max-height: 45px;
+}
